cleaned up join
